+++ sqe-database-access/Models/ArtefactModels.cs
@@ -11,8 +11,8 @@
         public uint ArtefactDataEditorId { get; set; }
         public string Mask { get; set; }
         public uint MaskEditorId { get; set; }
-        public float? Scale { get; set; }
-        public float? Rotate { get; set; }
+        public decimal? Scale { get; set; }
+        public decimal? Rotate { get; set; }
         public uint? TranslateX { get; set; }
         public uint? TranslateY { get; set; }
         public uint PositionEditorId { get; set; }
